pyproject.toml
[tool.poetry]
name = "restaurant_app"
version = "0.1.0"
description = ""
authors = ["Ruslan Albakov <albakov.ruslan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.34"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
uvloop = "^0.20.0"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
greenlet = "^3.1.1"
redis = "^5.1.1"
setuptools = "^75.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
